% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_ipeadata.R
\name{aux_ipeadata}
\alias{aux_ipeadata}
\title{A function to extract BACEN series using their API}
\usage{
aux_ipeadata(arg1, periodicity)
}
\arguments{
\item{arg1}{Bacen series number.}

\item{periodicity}{A string specifying the periodicity.}
}
\description{
A function to extract BACEN series using their API
}
\keyword{bacen}
