% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_sidra.R
\name{aux_sidra}
\alias{aux_sidra}
\title{A function to extract Sidra series using their API}
\usage{
aux_sidra(cod_tabela, periods, variable, territory)
}
\arguments{
\item{cod_tabela}{Sidra series number.}

\item{periods}{Specifies the desired periods levels. Defaults to 'n1/1' - Brazil.}

\item{variable}{Specifies the desired variable levels. Defaults to 'n1/1' - Brazil.}

\item{territory}{Specifies the desired territorial levels. Defaults to 'n1/1' - Brazil.}
}
\description{
The different parameters define a table and its dimensions (periods, variables, territorial units and classification) to be consulted. The parameters that define the territorial units can be used more than once, allowing multiple distinct territorial searches with one request. The parameters that define the classes may vary from table to table, with a maximum of 6. Henceforth, a Sidra table ranges between 3 and 9 differente dimension (3 mandatory - periodos, var and terr and 6 optional).
}
\keyword{sidra}
