% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplo.R
\docType{data}
\name{datas}
\alias{datas}
\title{A Cat Function}
\format{An object of class \code{character} of length 1.}
\usage{
datas
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
