% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoseries.R
\name{ecoseries}
\alias{ecoseries}
\title{An R interface to Brazilian Central Bank and Sidra APIs and the IPEA data}
\description{
The ecoseries package has three main functions, one to gather data from the Brazilian Central Bank (Bacen) and another to gather Institute of Applied Economic Research data (IPEA).

The series_bacen function takes advantage of the Bacen API ("http://api.bcb.gov.br/") to access this sistem "https://www3.bcb.gov.br/sgspub/localizarseries/localizarSeries.do?method=prepararTelaLocalizarSeries", load and save the series. You can either save in csv or xlsx formats.

The series_ipea function accesses a simpler section of the ipeadata website and harvests the informations based on series numbers.

The series_sidra function uses the Sidra API ("http://api.sidra.ibge.gov.br") to load and save the series. You can either save in csv or xlsx formats.
}
\keyword{bacen}
\keyword{bcb}
\keyword{economics}
\keyword{ibge}
\keyword{ipea}
\keyword{ipeadata}
\keyword{sidra}
