% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_bacen_anna.R
\name{series_bacen_anna}
\alias{series_bacen_anna}
\title{A Cat Function}
\usage{
series_bacen_anna(arg1, ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
