% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_bacen_luma.R
\name{series_bacen_luma}
\alias{series_bacen_luma}
\title{A function to extract BACEN series specific to LUMA}
\usage{
series_bacen_luma(arg1, ..., file = NULL, save = "csv")
}
\arguments{
\item{arg1}{Bacen series number.}

\item{...}{More series number.}

\item{file}{A xls or xlsx file you would like the function to open for you. Defaults to NULL.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}
}
\description{
A function to extract BACEN series specific to LUMA
}
\keyword{cats}
