% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_ipeadata.R
\name{series_ipeadata}
\alias{series_ipeadata}
\title{A function to extract IPEA series using their API}
\usage{
series_ipeadata(arg1, ..., periodicity = c("Y", "D"), save = "")
}
\arguments{
\item{arg1}{ipea series number.}

\item{...}{More series number.}

\item{periodicity}{A string specifying the periodicity.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}
}
\description{
A function to extract IPEA series using their API
}
\examples{
ipea=series_ipeadata(394221910, periodicity = c("Y")) 
}
\keyword{ipea}
\keyword{ipeadata}

