% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_ipeadata_old.R
\name{series_ipeadata_old}
\alias{series_ipeadata_old}
\title{A function to extract Sidra series using their API}
\usage{
series_ipeadata_old(arg1, ..., periodicity = "D", save = "")
}
\arguments{
\item{arg1}{ipea series number.}

\item{...}{More series number.}

\item{periodicity}{A string specifying the periodicity.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}
}
\description{
A function to extract Sidra series using their API
}
\keyword{ipea}
\keyword{ipeadata}
