% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_bacen.R
\name{series_bacen}
\alias{series_bacen}
\title{A function to extract BACEN series using their API}
\usage{
series_bacen(arg1, ..., save = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
series_bacen(1242,2134)
}
\keyword{cats}

