% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_bacen.R, R/series_bacen_luma.R
\name{series_bacen}
\alias{series_bacen}
\title{A function to extract BACEN series using their API}
\usage{
series_bacen(arg1, ..., file = NULL, save = "csv")

series_bacen(arg1, ..., file = NULL, save = "csv")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}

\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.

This function allows you to express your love of cats.
}
\examples{
series_bacen(1242,2134)
series_bacen(1242,2134, file = "C:/Users/fernando.teixeira/Dropbox/Luciana+Maclá/Índice de Atividade Econômica do Bacen (IBC-BR) - Com ajuste.xlsm")
series_bacen(1242,2134)
series_bacen(1242,2134, file = "C:/Users/fernando.teixeira/Dropbox/Luciana+Maclá/Índice de Atividade Econômica do Bacen (IBC-BR) - Com ajuste.xlsm")
}
\keyword{cats}

