% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_bacen.R
\name{series_bacen}
\alias{series_bacen}
\title{A function to extract BACEN series using their API}
\usage{
series_bacen(arg1, ..., save = "")
}
\arguments{
\item{arg1}{Bacen series number.}

\item{...}{More series number.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}
}
\description{
A function to extract BACEN series using their API
}
\examples{
series_bacen(1242,2134)
}
\keyword{bacen}

