% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_bacen.R
\name{series_bacen}
\alias{series_bacen}
\title{A function to extract BACEN series using their API}
\usage{
series_bacen(x, from = "", to = "", save = "")
}
\arguments{
\item{x}{Bacen series numbers. Either an integer or a numeric vector.}

\item{from}{A string specifying where the series shall start.}

\item{to}{A string specifying where the series shall end.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}
}
\description{
A function to extract BACEN series using their API
}
\examples{
bacen = series_bacen(x=c(2465,2134))
}
\keyword{bacen}
