% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_sidra.R
\name{series_sidra}
\alias{series_sidra}
\title{A function to extract Sidra series using their API}
\usage{
series_sidra(arg1, ..., save = "")
}
\arguments{
\item{arg1}{Sidra series number.}

\item{...}{More series number.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}
}
\description{
A function to extract Sidra series using their API
}
\keyword{sidra}
