% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_sidra.R
\name{series_sidra}
\alias{series_sidra}
\title{A function to extract Sidra series using their API}
\usage{
series_sidra(cod_tabela, periods = "all", variable = "all",
  territory = c(n1 = "brazil", n2 = "region", n3 = "state", n6 =
  "municipality"), classes = "", dec = 2, header = FALSE, save = "",
  form = "a")
}
\arguments{
\item{cod_tabela}{Sidra series number.}

\item{periods}{Specifies the desired periods. Defaults to 'all'.}

\item{variable}{Specifies the desired variables. Defaults to 'allxp'.}

\item{territory}{Specifies the desired territorial levels. Defaults to 'n1/1' - Brazil.}

\item{classes}{Specifies the table classifications. Defaults to "".}

\item{dec}{Specifies the desired decimals in table numbers.}

\item{header}{Logical. Either TRUE or FALSE.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}

\item{form}{Specifies the formatting of the table. Defaults to 'n' - metadata names only.}
}
\description{
The different parameters define a table and its dimensions (periods, variables, territorial units and classification) to be consulted. The parameters that define the territorial units can be used more than once, allowing multiple distinct territorial searches with one request. The parameters that define the classes may vary from table to table, with a maximum of 6. Henceforth, a Sidra table ranges between 3 and 9 differente dimension (3 mandatory - periods, var and terr and 6 optional).
}
\keyword{sidra}
