% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_sidra.R
\name{series_sidra}
\alias{series_sidra}
\title{A function to extract Sidra series using their API}
\usage{
series_sidra(x, from = NULL, to = NULL, territory = c(n1 = "brazil", n2 =
  "region", n3 = "state"), header = TRUE, save = "", variable = "allxp",
  sections = NULL)
}
\arguments{
\item{x}{Sidra series number.}

\item{from}{A string or character vector specifying where the series shall start. Defaults to 1980.}

\item{to}{A string or character vector specifying where the series shall end. Defaults to current year.}

\item{territory}{Specifies the desired territorial levels.}

\item{header}{Logical. Either TRUE or FALSE.}

\item{save}{A string specifying if data should be saved in csv or xlsx format. 
Defaults to not saving.}

\item{variable}{An integer describing what variable characteristics are to be returned. Defaults to all available.}

\item{sections}{A vector containing the classification code in the first slot and the desired tables from this classification.}
}
\description{
The different parameters define the table and its dimensions (periods, variables, territorial units and classification) to be consulted. The parameters that define the sections may vary from table to table. Henceforth, the Sidra function ranges between 2 mandatory arguments - x (the series number) and territory (the geographic scope) to 6 arguments, where you can input the time window wanted, the variables and the sections. You can only choose one variable per series per request, but multiple sections within the variable.
}
\examples{
sidra=series_sidra(x = c(1612), from = 1990, to = 2015, territory = "brazil")
sidra=series_sidra(x = c(3653), from = c("200201"), 
to = c("201512"), territory = "brazil", 
variable = 3135, sections = list(c(544,129316,129330)))
}
\keyword{sidra}
